Use BigDecimal for precise calculations in trade logic
Replaced Double with BigDecimal for better precision in lot size, tick size, and modifier value handling. Adjusted related calculations and ensured proper type conversion where necessary. This enhances numerical accuracy in critical trade operations.

+++ backend/src/main/kotlin/uk/co/threebugs/darwinexclient/modifiers/ModifierEntity.kt
@@ -1,6 +1,7 @@
 package uk.co.threebugs.darwinexclient.modifier
 
 import jakarta.persistence.*
+import java.math.BigDecimal
 
 @Entity
 @Table(name = "modifiers")
@@ -12,6 +13,6 @@ data class Modifier(
     @Column(nullable = false, length = 255)
     val name: String, // Name of the modifier
 
-    @Column(nullable = false, precision = 10, scale = 2)
-    val value: Double // Value of the modifier
+    @Column(nullable = false, precision = 10, scale = 2) // Precisely defines the scale for the value
+    val value: BigDecimal // Using BigDecimal for exact precision
 )
